refactor: EntityMetadataInspector 구조 변경 적용
