+++ src/test/java/persistence/JdbcTemplateTest.java
@@ -3,7 +3,6 @@ package persistence;
 import database.DatabaseServer;
 import database.H2;
 import jdbc.JdbcTemplate;
-import jdbc.RowMapper;
 import jdbc.RowMapperImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
